backend/requirements.txt
asgiref==3.8.1
Brotli==1.1.0
click==8.1.7
colorama==0.4.6
Django==4.2.13
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-writable-nested==0.7.2
gunicorn==23.0.0
h11==0.14.0
packaging==24.1
pillow==12.1.0
psycopg2-binary==2.9.10
PyJWT==2.8.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2025.3
